Move add command back for now.
